+++ app/database/migrations/2020_07_17_012659_add_breakdown_of_costs_to_invoices.php
@@ -13,12 +13,11 @@ class AddBreakdownOfCostsToInvoices extends Migration
     public function up()
     {
         Schema::table('users_invoices', function (Blueprint $table) {
-            //
-            $table->float('call_costs')->default(0.0);
-            $table->float('recording_costs')->default(0.0);
-            $table->float('fax_costs')->default(0.0);
-            $table->float('membership_costs')->default(0.0);
-            $table->float('number_costs')->default(0.0);
+            $table->integer('call_costs')->default(0);
+            $table->integer('recording_costs')->default(0);
+            $table->integer('fax_costs')->default(0);
+            $table->integer('membership_costs')->default(0);
+            $table->integer('number_costs')->default(0);
         });
     }
